--- a/hs/src/P14.hs
+++ b/hs/src/P14.hs
@@ -1,15 +1,18 @@
 module P14 where
 
+import Data.Attoparsec.Text (anyChar, count)
 import qualified Data.Map as Map
 import Import hiding (mask)
 
 type Input = [MaskGroup]
 
-type MaskGroup = (Mask, [(Int, Int)])
+type MaskGroup = (Mask, [(Int64, Int64)])
 
-type Mask = [(Int, Maybe Bit)]
+type Mask = [(Int, Char)]
 
-type RAM = Map Int Int
+type RAM = Map Int64 Int64
+
+type Strategy = Mask -> RAM -> (Int64, Int64) -> RAM
 
 data Bit = O | I deriving (Show, Eq)
 
@@ -18,29 +21,29 @@ solution = solve parser $ \input -> do
   part1 $ sum $ process v1 input -- 12408060320841
   part2 $ sum $ process v2 input -- 4466434626828
 
-process :: (Mask -> RAM -> (Int, Int) -> RAM) -> Input -> RAM
-process apply = foldl processGroup Map.empty
+process :: Strategy -> Input -> RAM
+process apply = foldl' processGroup Map.empty
   where
-    processGroup acc (mask, sets) = foldl (apply mask) acc sets
+    processGroup acc (mask, sets) = foldl' (apply mask) acc sets
 
-v1 :: Mask -> RAM -> (Int, Int) -> RAM
+v1 :: Strategy
 v1 mask ram (addr, value) = Map.insert addr (masked mask value) ram
 
-masked :: Mask -> Int -> Int
-masked mask n = foldr bitmask n mask
+masked :: Mask -> Int64 -> Int64
+masked mask n = foldl' bitmask n mask
   where
-    bitmask (i, Just O) = flip clearBit i
-    bitmask (i, Just I) = flip setBit i
-    bitmask _ = identity
+    bitmask acc (i, '0') = acc `clearBit` i
+    bitmask acc (i, '1') = acc `setBit` i
+    bitmask acc _ = acc
 
-v2 :: Mask -> RAM -> (Int, Int) -> RAM
-v2 mask ram (addr, value) = foldl (\r i -> Map.insert i value r) ram $ floats mask addr
+v2 :: Strategy
+v2 mask ram (addr, value) = Map.fromList [(key, value) | key <- floats mask addr] <> ram
 
-floats :: Mask -> Int -> [Int]
+floats :: Mask -> Int64 -> [Int64]
 floats mask base = foldM f base mask
   where
-    f n (i, Nothing) = [n `setBit` i, n `clearBit` i]
-    f n (i, Just I) = [n `setBit` i]
+    f n (i, 'X') = [n `setBit` i, n `clearBit` i]
+    f n (i, '1') = [n `setBit` i]
     f n _ = [n]
 
 parser :: Parser Input
@@ -51,14 +54,7 @@ parser = commands `sepBy` "\n"
         <$> ("mask = " *> mask <* "\n")
         <*> set `sepBy` "\n"
 
-    mask =
-      fmap (zip [width - 1, width - 2 .. 0]) $
-        many $
-          choice
-            [ "X" $> Nothing,
-              "0" $> Just O,
-              "1" $> Just I
-            ]
+    mask = zip [width - 1, width - 2 .. 0] <$> count width anyChar
 
     set =
       (,)