∆ 265.5 ms to 136.9 ms on my machine.

    time                 265.5 ms   (228.1 ms .. 320.4 ms)
                         0.984 R²   (0.971 R² .. 1.000 R²)
    mean                 292.4 ms   (272.4 ms .. 328.7 ms)
    std dev              38.48 ms   (6.766 ms .. 53.42 ms)

to

    time                 136.9 ms   (120.9 ms .. 155.8 ms)
                         0.979 R²   (0.923 R² .. 1.000 R²)
    mean                 142.1 ms   (137.3 ms .. 151.2 ms)
    std dev              9.179 ms   (4.818 ms .. 13.12 ms)

There's still some work to do here. 50-60% of the time and alloc'ing is in `floats` and it seems like there's room for improvement there.